requirements.txt
Flask==1.1.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
